feat: 수익이 반전되는 기능 구현
