+++ src/test/java/blackjack/domain/betting/ProfitTest.java
@@ -1,9 +1,16 @@
 package blackjack.domain.betting;
 
-import org.assertj.core.api.Assertions;
+import static org.assertj.core.api.Assertions.assertThat;
+
+import org.junit.jupiter.api.DisplayNameGeneration;
+import org.junit.jupiter.api.DisplayNameGenerator;
+import org.junit.jupiter.api.Nested;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
+@SuppressWarnings("NonAsciiCharacters")
+@DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 class ProfitTest {
 
     @ParameterizedTest(name = "원금: {0}, 결과: {1}")
@@ -11,25 +18,30 @@ class ProfitTest {
     void 수익이_50_퍼센트_증가한다(final int value, final int expected) {
         final Profit profit = new Profit(value);
 
-        final Profit increaseProfit = profit.increaseFiftyPercent();
+        final Profit increase = profit.increaseFiftyPercent();
 
-        Assertions.assertThat(increaseProfit.getValue()).isEqualTo(expected);
-    }
-}
-
-class Profit {
-
-    private final int value;
-
-    public Profit(final int value) {
-        this.value = value;
+        assertThat(increase.getValue()).isEqualTo(expected);
     }
 
-    public Profit increaseFiftyPercent() {
-        return new Profit(value + (int) (value * 0.5));
-    }
+    @Nested
+    class reverse_메서드는 {
 
-    public int getValue() {
-        return value;
+        @Test
+        void 수익이_양수인_경우_음수가_된다() {
+            final Profit profit = new Profit(10000);
+
+            final Profit reverse = profit.reverse();
+
+            assertThat(reverse.getValue()).isEqualTo(-10000);
+        }
+
+        @Test
+        void 수익이_음수인_경우_양수가_된다() {
+            final Profit profit = new Profit(-10000);
+
+            final Profit reverse = profit.reverse();
+
+            assertThat(reverse.getValue()).isEqualTo(10000);
+        }
     }
 }
