+++ src/test/java/blackjack/domain/participant/ResultTest.java
@@ -5,9 +5,13 @@ import static blackjack.domain.participant.Result.LOSE;
 import static blackjack.domain.participant.Result.WIN;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import org.junit.jupiter.api.DisplayNameGeneration;
+import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
+@SuppressWarnings("NonAsciiCharacters")
+@DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 class ResultTest {
 
     @Nested
